requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10  # PostgreSQL driver

# Security
pyjwt==2.9.0

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0

